requirements.txt
asgiref==3.9.2
click==8.3.0
Django==5.2.7
django-debug-toolbar==6.0.0
djangorestframework==3.14.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
gunicorn==23.0.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
whitenoise==6.7.0
